265 - ms to ns in vehicle-os: implemented feedback (int->long), changed x*1000*ms -> x*s

+++ app/tutorials/example-applications/src/main/java/org/eclipse/mosaic/app/tutorial/configurableapp/VehicleConfigurationApp.java
@@ -69,7 +69,7 @@ public class VehicleConfigurationApp extends ConfigurableApplication<CExample, V
             int wantedSpeed = this.config.fooInteger;
             if (getOs().getVehicleData().getSpeed() > wantedSpeed) {
                 getLog().info("The current speed equals {}m/s will be slowed down to the wanted speed equals {}m/s", getOs().getVehicleData().getSpeed(), wantedSpeed);
-                getOs().slowDown(wantedSpeed, (int) (10 * TIME.SECOND));
+                getOs().slowDown(wantedSpeed, 10 * TIME.SECOND);
             }
         } else {
             getLog().warn("Configuration wasn't loaded from json file!");
